requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
psycopg2-binary==2.9.9

# Auth
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0

# Payments
stripe==10.10.0

# HTTP client (Supabase Storage)
httpx==0.27.2
python-dateutil==2.9.0

# Env vars
python-dotenv==1.0.1
